+++ src/main/java/com/ntl/trello/repository/UsersRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersRepository extends CrudRepository<Users, Long> {
+public interface UsersRepository extends CrudRepository<Users, Integer> {
     Users findByUsername(String username);
     Users findById(int id);
 }
